feat: add toAdminResponse function to map DukanProduct to DukanProductAdminResponse
