+++ dukan/src/main/kotlin/net/thechance/dukan/api/mapper/product/DukanProductMapper.kt
@@ -1,6 +1,8 @@
 package net.thechance.dukan.api.mapper.product
 
+import net.thechance.dukan.api.dto.product.DukanProductAdminResponse
 import net.thechance.dukan.api.dto.product.DukanProductResponse
+import net.thechance.dukan.entity.DukanProduct
 import net.thechance.dukan.service.model.DukanProductWithFavoriteAndQuantity
 
 
@@ -16,3 +18,13 @@ fun DukanProductWithFavoriteAndQuantity.toResponse() = DukanProductResponse(
     isFavorite = this.isFavorite,
     dukanId = this.product.dukan.id
     )
+
+fun DukanProduct.toAdminResponse() = DukanProductAdminResponse(
+    id = this.id,
+    name = this.name,
+    price = this.price,
+    discountedPrice = null, //TODO
+    description = this.description,
+    imageUrls = this.imageUrls,
+    createdAt = this.createdAt,
+)
